requirements.txt
feedparser
PyYAML
python-frontmatter
python-dateutil